math/tinystdio/stdlib: Avoid warnings when long double == double
Even if long double and double are the same underlying machine type,
gcc cares if you mix them without suitable casts. Disable some
warnings for pure function aliases and add some casts in other places.

+++ newlib/libm/common/fabsl.c
@@ -36,7 +36,7 @@ POSSIBILITY OF SUCH DAMAGE.
 long double
 fabsl (long double x)
 {
-  return fabs(x);
+  return (long double) fabs((double) x);
 }
 #endif
